+++ logwise.go
@@ -26,54 +26,24 @@ package logwise
 
 import (
 	"fmt"
+	"github.com/logwise/loglevel"
 	"io"
 	"log"
 	"os"
 	"sync"
 )
 
-const (
-	TraceLevel   LogLevel = 0
-	DebugLevel   LogLevel = 10
-	InfoLevel    LogLevel = 20
-	WarningLevel LogLevel = 30
-	ErrorLevel   LogLevel = 40
-	FatalLevel   LogLevel = 50
-	SystemLevel  LogLevel = 100
-)
-
 // 默认 flag
 const DefaultFlags = log.Ldate | log.Ltime | log.Lshortfile
 
-// 日志过滤等级
-var level LogLevel
-
 // 默认输出
 var DefaultOutput = os.Stdout
-
-type LogLevel int
-
-func (l LogLevel) String() string {
-	var name string
-	switch l {
-	case TraceLevel:
-		name = "TRACE"
-	case DebugLevel:
-		name = "DEBUG"
-	case InfoLevel:
-		name = "INFO"
-	case WarningLevel:
-		name = "WARNING"
-	case ErrorLevel:
-		name = "ERROR"
-	}
-	return name
-}
 
 type Logger struct {
 	prefix string
 	out    io.Writer
 	mu     sync.Mutex
+	level  loglevel.LogLevel
 	logger *log.Logger
 }
 
@@ -83,6 +53,16 @@ func (l Logger) String() string {
 
 func (l Logger) Prefix() string {
 	return l.prefix
+}
+
+func (l Logger) FullPrefix(lv loglevel.LogLevel) string {
+	var ret string
+	if l.prefix != "" {
+		ret = l.prefix + " " + lv.GetPrefix()
+	} else {
+		ret = lv.GetPrefix()
+	}
+	return ret
 }
 
 func (l *Logger) SetPrefix(prefix string) {
@@ -97,11 +77,19 @@ func (l *Logger) SetOutput(out io.Writer) {
 	l.out = out
 }
 
+func (l Logger) LogLevel() loglevel.LogLevel {
+	return l.level
+}
+
+func (l *Logger) SetLogLevel(lv loglevel.LogLevel) {
+	l.level = lv
+}
+
 // 按照不同的 loglevel 改变 logger 属性
-func (l *Logger) dress(lv LogLevel) {
-	l.logger.SetPrefix(GetPrefix(lv, l.prefix))
+func (l *Logger) dress(lv loglevel.LogLevel) {
+	l.logger.SetPrefix(l.FullPrefix(lv))
 	if l.out == DefaultOutput {
-		if lv != SystemLevel && lv >= ErrorLevel {
+		if lv != loglevel.System && lv >= loglevel.Error {
 			l.logger.SetOutput(os.Stderr)
 		} else {
 			l.logger.SetOutput(l.out)
@@ -111,8 +99,8 @@ func (l *Logger) dress(lv LogLevel) {
 	}
 }
 
-func (l *Logger) print(lv LogLevel, v ...interface{}) {
-	if lv == SystemLevel || lv >= level {
+func (l *Logger) print(lv loglevel.LogLevel, v ...interface{}) {
+	if lv == loglevel.System || lv >= l.level {
 		l.mu.Lock()
 		l.dress(lv)
 		_ = l.logger.Output(3, fmt.Sprint(v...))
@@ -120,8 +108,8 @@ func (l *Logger) print(lv LogLevel, v ...interface{}) {
 	}
 }
 
-func (l *Logger) println(lv LogLevel, v ...interface{}) {
-	if lv == SystemLevel || lv >= level {
+func (l *Logger) println(lv loglevel.LogLevel, v ...interface{}) {
+	if lv == loglevel.System || lv >= l.level {
 		l.mu.Lock()
 		l.dress(lv)
 		_ = l.logger.Output(3, fmt.Sprintln(v...))
@@ -130,8 +118,8 @@ func (l *Logger) println(lv LogLevel, v ...interface{}) {
 
 }
 
-func (l *Logger) printf(lv LogLevel, format string, v ...interface{}) {
-	if lv == SystemLevel || lv >= level {
+func (l *Logger) printf(lv loglevel.LogLevel, format string, v ...interface{}) {
+	if lv == loglevel.System || lv >= l.level {
 		l.mu.Lock()
 		l.dress(lv)
 		_ = l.logger.Output(3, fmt.Sprintf(format, v...))
@@ -139,108 +127,121 @@ func (l *Logger) printf(lv LogLevel, format string, v ...interface{}) {
 	}
 }
 
-// Default, based on global log level
+// Default, based on the log level of this logger
 func (l *Logger) Print(v ...interface{}) {
-	l.print(level, v...)
+	l.print(l.level, v...)
 }
 
 func (l *Logger) Println(v ...interface{}) {
-	l.println(level, v...)
+	l.println(l.level, v...)
 }
 
 func (l *Logger) Printf(format string, v ...interface{}) {
-	l.printf(level, format, v...)
+	l.printf(l.level, format, v...)
+}
+
+// NOTSET
+func (l *Logger) NotSet(v ...interface{}) {
+	l.print(loglevel.NotSet, v...)
+}
+
+func (l *Logger) NotSetln(v ...interface{}) {
+	l.println(loglevel.NotSet, v...)
+}
+
+func (l *Logger) NotSetf(format string, v ...interface{}) {
+	l.printf(loglevel.NotSet, format, v...)
 }
 
 // TRACE
 func (l *Logger) Trace(v ...interface{}) {
-	l.print(TraceLevel, v...)
+	l.print(loglevel.Trace, v...)
 }
 
 func (l *Logger) Traceln(v ...interface{}) {
-	l.println(TraceLevel, v...)
+	l.println(loglevel.Trace, v...)
 }
 
 func (l *Logger) Tracef(format string, v ...interface{}) {
-	l.printf(TraceLevel, format, v...)
+	l.printf(loglevel.Trace, format, v...)
 }
 
 // DEBUG
 func (l *Logger) Debug(v ...interface{}) {
-	l.print(DebugLevel, v...)
+	l.print(loglevel.Debug, v...)
 }
 
 func (l *Logger) Debugln(v ...interface{}) {
-	l.println(DebugLevel, v...)
+	l.println(loglevel.Debug, v...)
 }
 
 func (l *Logger) Debugf(format string, v ...interface{}) {
-	l.printf(DebugLevel, format, v...)
+	l.printf(loglevel.Debug, format, v...)
 }
 
 // INFO
 func (l *Logger) Info(v ...interface{}) {
-	l.print(InfoLevel, v...)
+	l.print(loglevel.Info, v...)
 }
 
 func (l *Logger) Infoln(v ...interface{}) {
-	l.println(InfoLevel, v...)
+	l.println(loglevel.Info, v...)
 }
 
 func (l *Logger) Infof(format string, v ...interface{}) {
-	l.printf(InfoLevel, format, v...)
+	l.printf(loglevel.Info, format, v...)
 }
 
 // WARNING
 func (l *Logger) Warn(v ...interface{}) {
-	l.print(WarningLevel, v...)
+	l.print(loglevel.Warning, v...)
 }
 
 func (l *Logger) Warnln(v ...interface{}) {
-	l.println(WarningLevel, v...)
+	l.println(loglevel.Warning, v...)
 }
 
 func (l *Logger) Warnf(format string, v ...interface{}) {
-	l.printf(WarningLevel, format, v...)
+	l.printf(loglevel.Warning, format, v...)
 }
 
 // ERROR
 func (l *Logger) Error(v ...interface{}) {
-	l.print(ErrorLevel, v...)
+	l.print(loglevel.Error, v...)
 }
 
 func (l *Logger) Errorln(v ...interface{}) {
-	l.println(ErrorLevel, v...)
+	l.println(loglevel.Error, v...)
 }
 
 func (l *Logger) Errorf(format string, v ...interface{}) {
-	l.printf(ErrorLevel, format, v...)
+	l.printf(loglevel.Error, format, v...)
 }
 
 // FATAL
 func (l *Logger) Fatal(v ...interface{}) {
-	l.print(FatalLevel, v...)
+	l.print(loglevel.Fatal, v...)
 }
 
 func (l *Logger) Fatalln(v ...interface{}) {
-	l.println(FatalLevel, v...)
+	l.println(loglevel.Fatal, v...)
 }
 
 func (l *Logger) Fatalf(format string, v ...interface{}) {
-	l.printf(FatalLevel, format, v...)
+	l.printf(loglevel.Fatal, format, v...)
 }
 
 // SYSTEM
 func (l *Logger) System(v ...interface{}) {
-	l.print(SystemLevel, v...)
+	l.print(loglevel.System, v...)
 }
 
 func (l *Logger) Systemln(v ...interface{}) {
-	l.println(SystemLevel, v...)
+	l.println(loglevel.System, v...)
 }
 
 func (l *Logger) Systemf(format string, v ...interface{}) {
-	l.printf(SystemLevel, format, v...)
+	l.printf(loglevel.System, format, v...)
 }
 
 // New logger
@@ -252,61 +253,7 @@ func New(out io.Writer, prefix string, flag int) *Logger {
 	return logger
 }
 
-// 默认 logger
+// Default logger
 func Default() *Logger {
 	return New(DefaultOutput, "", DefaultFlags)
-}
-
-// 获取当前的 loglevel
-func GetLogLevel() LogLevel {
-	return level
-}
-
-// 设置当前的 loglevel
-func SetLogLevel(l LogLevel) {
-	level = l
-}
-
-// 转换成 LogLevel
-func ConvertStringToLogLevel(logLevel string) LogLevel {
-	var level LogLevel
-	switch logLevel {
-	case "TRACE":
-		level = TraceLevel
-	case "DEBUG":
-		level = DebugLevel
-	case "INFO":
-		level = InfoLevel
-	case "WARNING":
-		level = WarningLevel
-	case "ERROR":
-		level = ErrorLevel
-	case "FATAL":
-		level = FatalLevel
-	}
-	return level
-}
-
-// 获取 Logger 前缀
-func GetPrefix(level LogLevel, prefix string) string {
-	if prefix != "" {
-		prefix += " "
-	}
-	switch level {
-	case TraceLevel:
-		prefix += "[TRACE]  "
-	case DebugLevel:
-		prefix += "[DEBUG]  "
-	case InfoLevel:
-		prefix += "[INFO]   "
-	case WarningLevel:
-		prefix += "[WARN]   "
-	case ErrorLevel:
-		prefix += "[ERROR]  "
-	case FatalLevel:
-		prefix += "[FATAL]  "
-	case SystemLevel:
-		prefix += "[SYS]    "
-	}
-	return prefix
 }